refactor: clean up code and fix bugs
License: MIT
