+++ package.json
@@ -13,13 +13,13 @@
   },
   "dependencies": {
     "@web3-storage/data-segment": "^3.0.1",
+    "chalk": "^5.3.0",
     "cli-progress": "^3.12.0",
     "fr32-sha2-256-trunc254-padded-binary-tree-multihash": "^1.1.4",
     "p-map": "^6.0.0",
     "p-retry": "^6.1.0",
     "piscina": "^4.1.0",
-    "sade": "^1.8.1",
-    "undici": "^5.25.2"
+    "sade": "^1.8.1"
   },
   "devDependencies": {
     "@tsconfig/node18": "^18.2.2",
