+++ src/worker.js
@@ -1,8 +1,11 @@
 import { pipeline } from 'node:stream/promises'
 import { createReadStream } from 'node:fs'
-import { PieceHash } from './index.js'
+import { PieceHash } from './piece.js'
 
 /**
+ * Return the piece cid string for a given fs path to a car
+ * The entry point for worker threads created by `piece verify`
+ *
  * @param {string} car
  */
 export default async function worker (car) {
@@ -13,7 +16,7 @@ export default async function worker (car) {
 /**
  * @param {string} car - path to car
  */
-async function pieceCidForCar (car) {
+export async function pieceCidForCar (car) {
   const pieceHash = new PieceHash()
   await pipeline(
     createReadStream(car),
